Main virtual learn enviroment
In the main computer, Kali AI will be taught how to use its own tools and how to think as the greatest experts would think to crash systems after it will generate a cybersecurity report that can lead to improvements in security. Furthermore, in the future, I will create the Kali Security builder that will use the data of KAI to create an unbreakable firewall. Also, probably I will create a Crypto token to finance the project and share security power with the community.